requirements.txt
selenium>=4.15.0
pyautogui>=0.9.54
pyttsx3>=2.90
SpeechRecognition>=3.10.0
python-dotenv>=1.0.0
pinecone>=3.0.0
Pillow>=10.0.0
keyboard>=0.13.5
pynput>=1.7.6
pytest>=7.4.0
google-generativeai>=0.8.0
rich>=13.0.0
prompt_toolkit>=3.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
jinja2>=3.1.0
# Screen control module dependencies
opencv-python>=4.8.0
pygetwindow>=0.0.9
pyperclip>=1.8.2
numpy>=1.24.0
pywin32>=306

# ── Browser control module dependencies ───────────────────────────────────────
# After pip install -r requirements.txt, also run:
#   playwright install chromium
playwright>=1.44.0

# ── Google API integrations (Gmail, Calendar, Docs) ───────────────────────────
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
